# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/inventory/levels": {
      "get": {
        "operationId": "InventoryController_levels",
        "parameters": [
          {
            "name": "warehouse_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "item_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/inventory/movements": {
      "get": {
        "operationId": "InventoryController_movements",
        "parameters": [
          {
            "name": "item_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "warehouse_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/inventory/reorder": {
      "get": {
        "operationId": "InventoryController_reorder",
        "parameters": [
          {
            "name": "warehouse_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
...
    "/api/v1/inventory/valuation/as-of": {
      "get": {
        "operationId": "InventoryController_valuationAsOf",
        "parameters": [
          {
            "name": "as_of",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "warehouse_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "item_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/inventory/balances/recompute": {
      "post": {
        "operationId": "InventoryController_recompute",
        "parameters": [
          {
            "name": "warehouse_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "item_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/inventory/transfers": {
      "get": {
        "operationId": "InventoryController_listTransfers",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      },
      "post": {
        "operationId": "InventoryController_transfer",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/inventory/transfers/{id}": {
      "get": {
        "operationId": "InventoryController_getTransfer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
...
    "/api/v1/inventory/transfers/{id}/send": {
      "post": {
        "operationId": "InventoryController_sendTransfer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/inventory/transfers/{id}/receive": {
      "post": {
        "operationId": "InventoryController_receiveTransfer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
...
    "/api/v1/inventory/transfers/{id}/cancel": {
      "post": {
        "operationId": "InventoryController_cancelTransfer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
...
    "/api/v1/inventory/lots": {
      "get": {
        "operationId": "InventoryController_lots",
        "parameters": [
          {
            "name": "item_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "warehouse_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      },
      "post": {
        "operationId": "InventoryController_createLot",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/inventory/expiry": {
      "get": {
        "operationId": "InventoryController_expiry",
        "parameters": [
          {
            "name": "warehouse_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/inventory/serials": {
      "get": {
        "operationId": "InventoryController_serials",
        "parameters": [
          {
            "name": "item_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "warehouse_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      },
      "post": {
        "operationId": "InventoryController_createSerial",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
...
    "/api/v1/inventory/documents": {
      "get": {
        "operationId": "InventoryDocumentsController_list",
        "parameters": [
          {
            "name": "kind",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "warehouse_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "InventoryDocuments"
        ]
      },
      "post": {
        "operationId": "InventoryDocumentsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "InventoryDocuments"
        ]
      }
    },
    "/api/v1/inventory/documents/{id}": {
      "get": {
        "operationId": "InventoryDocumentsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "InventoryDocuments"
        ]
      },
      "patch": {
        "operationId": "InventoryDocumentsController_updateDraft",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "InventoryDocuments"
        ]
      }
    },
    "/api/v1/inventory/documents/{id}/post": {
      "post": {
        "operationId": "InventoryDocumentsController_post",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "InventoryDocuments"
        ]
      }
    },
    "/api/v1/inventory/documents/{id}/cancel": {
      "post": {
        "operationId": "InventoryDocumentsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "InventoryDocuments"
        ]
      }
    },
    "/api/v1/inventory/documents/{id}/void": {
      "post": {
        "operationId": "InventoryDocumentsController_void",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "InventoryDocuments"
        ]
      }
    },
...
    "/api/v1/reports/print/inventory-documents/{id}": {
      "get": {
        "operationId": "ReportingController_inventoryDocumentPrint",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reporting"
        ]
      }
    },
...
  },
  "info": {
    "title": "ERP SaaS API",
    "description": "Cloud multi-tenant ERP. Errors are RFC 9457 application/problem+json with a stable `code` from @erp/contracts. Tenant context always comes from the access token.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "branchScope": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Branch-Id"
      },
      "idempotencyKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Idempotency-Key"
      }
    },
...
  }
}